% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_scenario.R
\name{ab_save}
\alias{ab_save}
\title{Save OD data as JSON files for import into A/B Street}
\usage{
ab_save(x, f)
}
\arguments{
\item{x}{A list object produced by \code{\link[=ab_scenario]{ab_scenario()}}}

\item{f}{A filename, e.g. \code{new_scenario.json}}
}
\value{
A JSON file containing scenarios from ab_scenario()
}
\description{
Save OD data as JSON files for import into A/B Street
}
