% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_scenario.R
\name{ab_sf}
\alias{ab_sf}
\title{Convert JSON representation of trips from A/B Street into an 'sf' object}
\usage{
ab_sf(json_df)
}
\arguments{
\item{json_df}{Character string or list representing a JSON file or list that
has been read into R and converted to a data frame.}
}
\value{
An \code{sf} data frame
}
\description{
This function takes a path to a JSON file representing an A/B Street
scenario, or an R representation of the JSON in a list, and returns
an \code{sf} object with the same structure as objects returned by
\code{\link[=ab_scenario]{ab_scenario()}}.
}
\details{
Note: the departure time in seconds is divided by 10000 on conversion
to represent seconds, which are easier to work with that 10,000th of
a second units.
}
\examples{
file_name = system.file("extdata/minimal_scenario.json", package = "abstr")
json_df = jsonlite::read_json(file_name, simplifyVector = TRUE)
ab_sf(json_df)
ab_sf(file_name)
json_df = jsonlite::read_json("inst/extdata/minimal_scenario2.json", simplifyVector = TRUE)
}
