% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_scenario.R
\name{ab_sf_to_json}
\alias{ab_sf_to_json}
\title{Convert geographic ('sf') representation of OD data to 'JSON list' structure}
\usage{
ab_sf_to_json(desire_lines_out, mode_column = "mode_base")
}
\arguments{
\item{desire_lines_out}{OD data represented as geographic lines created by
[ab_scenario()].}

\item{mode_column}{The column name in the desire lines data that contains
the mode of transport. `"mode_baseline"` by default.}
}
\value{
A list that can be saved as a JSON file with [ab_save()]
}
\description{
This function takes outputs from [ab_scenario()] and returns a list that
can be saved as a JSON file for import into A/B Street.
}
\examples{
library(sf) # for subsetting lines
ablines = ab_scenario(
  leeds_houses,
  leeds_buildings,
  leeds_desire_lines[7, ],
  leeds_zones,
  output_format = "sf"
)
ab_list = ab_sf_to_json(ablines, mode_column = "mode_base")
ab_list$scenario
ab_list$people$trips[[1]]
ablines_dutch = ab_scenario(
  leeds_houses,
  leeds_buildings,
  leeds_desire_lines[7, ],
  leeds_zones,
  scenario = "godutch",
  output_format = "sf"
)
ab_list = ab_sf_to_json(ablines_dutch, mode_column = "mode_godutch")
ab_list$scenario
ab_list$people$trips[[9]]
}
