% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{leeds_site_area}
\alias{leeds_site_area}
\alias{leeds_houses}
\alias{leeds_buildings}
\alias{leeds_desire_lines}
\alias{leeds_zones}
\alias{leeds_od}
\alias{sao_paulo_activity_df_20}
\alias{sao_paulo_activity_df_2}
\title{Datasets from Leeds and Sao Paulo}
\description{
These datasets represent OD data from Leeds and Sao Paulo.
}
