% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montlake_od.R
\docType{data}
\name{montlake_od}
\alias{montlake_od}
\title{Example OD Table for Montlake}
\format{
A data frame with columns:
\describe{
\item{o_id}{Trip origin zone ID (must match an ID in \code{montlake_zone}).}
\item{d_id}{Trip destination zone ID (must match an ID in \code{montlake_zone}).}
\item{Drive}{Count of how many trips were made using cars.}
\item{Transit}{Count of how many trips were made using public transit.}
\item{Bike}{Count of how many trips were made using bikes.}
\item{Walk}{Count of how many trips were made on foot.}
}
}
\usage{
montlake_od
}
\description{
Each row of this table contains a count of how many trips started in
column \code{o_id} and ended in column \code{d_id} according to different modes.
This example table has modes that match what A/B Street currently uses:
"Drive", "Transit", "Bike", and "Walk".
}
\details{
See the code used to create this data in "data-raw/montlake-test-data.R"
}
\keyword{datasets}
