% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montlake_zones.R
\docType{data}
\name{montlake_zones}
\alias{montlake_zones}
\title{Example Zones Table for Montlake}
\format{
A sf dataframe with columns:
\describe{
\item{id}{Zone ID (must match \code{o_id} or \code{d_id} in \code{montlake_od}).}
\item{geometry}{Simple feature collection (sfc) contain multipolygons,
each representing the boundaries of a TAZ near Montlake.}
}
}
\usage{
montlake_zones
}
\description{
Each row of this table contains the zone ID and geometry for a TAZ near
Montlake. Includes all zones that start or end within the Montlake
boundary and have at least 25 trips across all modes of transit in
\code{montlake_od}.
}
\details{
See the code used to create this data in "data-raw/montlake-test-data.R"
}
\keyword{datasets}
