% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sao_paulo_activity_df_20.R
\docType{data}
\name{sao_paulo_activity_df_20}
\alias{sao_paulo_activity_df_20}
\alias{sao_paulo_activity_df_2}
\title{Example Activity data for São Paulo}
\format{
A data frame with columns:
\describe{
\item{ID_PESS}{Person identifier.}
\item{CO_O_X}{Origin coordinate X.}
\item{CO_O_Y}{Origin coordinate Y.}
\item{CO_D_X}{Destination coordinate X.}
\item{CO_D_Y}{Destination coordinate Y.}
\item{MODOPRIN}{Main mode.}
\item{H_SAIDA}{Departure hour.}
\item{MIN_SAIDA}{Departure minute.}
}
}
\usage{
sao_paulo_activity_df_20
}
\description{
Each row of this table contains a single trip of people in the São Paulo city.
\code{sao_paulo_activity_df_2} represents the movement of 2 people,
\code{sao_paulo_activity_df_20} represents the movement of 20 people.
}
\details{
See the code used to create this data, and the full open dataset with
128 variables, the file
\href{https://github.com/a-b-street/abstr/blob/main/data-raw/sao-paulo-activity-data.R}{data-raw/sao-paulo-activity-data.R}
in the package's GitHub repo.
}
\examples{
dim(sao_paulo_activity_df_20)
names(sao_paulo_activity_df_20)
head(sao_paulo_activity_df_20)
dim(sao_paulo_activity_df_2)
names(sao_paulo_activity_df_2)
head(sao_paulo_activity_df_2)
}
\keyword{datasets}
