% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sao_paulo_activity_df_20.R
\docType{data}
\name{sao_paulo_activity_sf_20}
\alias{sao_paulo_activity_sf_20}
\alias{sao_paulo_activity_sf_2}
\title{Example Activity data for São Paulo}
\format{
A data frame with columns:
\describe{
\item{person}{Person identifier.}
\item{departure}{Departure time in seconds past midnight}
\item{mode}{Mode of travel in A/B Street terms}
\item{purpose}{Purpose of travel in A/B Street terms}
\item{geometry}{Geometry of the linestring representing the OD pair}
}
}
\usage{
sao_paulo_activity_sf_20
}
\description{
Each row of this table contains a single trip of people in the São Paulo city.
\code{sao_paulo_activity_sf_2} represents the movement of 2 people,
\code{sao_paulo_activity_sf_20} represents the movement of 20 people.
}
\details{
See the code used to create this data, and the full open dataset with
128 variables, the file
\href{https://github.com/a-b-street/abstr/blob/main/data-raw/sao-paulo-activity-data.R}{data-raw/sao-paulo-activity-data.R}
in the package's GitHub repo.
}
\examples{
dim(sao_paulo_activity_sf_20)
names(sao_paulo_activity_sf_20)
head(sao_paulo_activity_sf_20)
table(sao_paulo_activity_sf_20$mode)
table(sao_paulo_activity_sf_20$purpose)
dim(sao_paulo_activity_sf_2)
names(sao_paulo_activity_sf_2)
head(sao_paulo_activity_sf_2)
}
\keyword{datasets}
