% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_scenario.R
\name{ab_scenario}
\alias{ab_scenario}
\title{Generate A/B Street Scenario files/objects}
\usage{
ab_scenario(
  houses,
  buildings,
  desire_lines,
  zones,
  output_format = "json_list"
)
}
\description{
Generate A/B Street Scenario files/objects
}
\examples{
ab_scenario(
  leeds_houses,
  leeds_buildings,
  leeds_desire_lines,
  leeds_desire_zones,
  output_format = "sf"
)
}
