% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_scenario.R
\name{ab_scenario}
\alias{ab_scenario}
\title{Generate A/B Street Scenario files/objects}
\usage{
ab_scenario(
  houses,
  buildings,
  desire_lines,
  zones,
  output_format = "json_list"
)
}
\description{
Generate A/B Street Scenario files/objects
}
\examples{
dslines = leeds_desire_lines
ablines = ab_scenario(
  leeds_houses,
  leeds_buildings,
  dslines,
  leeds_zones,
  output_format = "sf"
)
plot(dslines$geometry, lwd = dslines[[3]] / 30)
plot(leeds_site_area$geometry, add = TRUE)
plot(leeds_buildings$geometry, add = TRUE)
plot(ablines$geometry, col = "blue", add = TRUE)
}
