% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_scenario.R
\name{ab_sf_to_json}
\alias{ab_sf_to_json}
\title{Convert geographic ('sf') representation of OD data to 'JSON list' structure}
\usage{
ab_sf_to_json(
  desire_lines_out,
  mode_column = "mode_base",
  time_fun = ab_time_normal,
  ...
)
}
\arguments{
\item{desire_lines_out}{OD data represented as geographic lines created by
[ab_scenario()].}

\item{mode_column}{The column name in the desire lines data that contains
the mode of transport. `"mode_baseline"` by default.}

\item{time_fun}{The function used to calculate departure times.
`ab_time_normal()` by default.}

\item{...}{Additional arguments to pass to `time_fun`}
}
\value{
A list that can be saved as a JSON file with [ab_save()]
}
\description{
This function takes outputs from [ab_scenario()] and returns a list that
can be saved as a JSON file for import into A/B Street.
}
\examples{
ablines = ab_scenario(
  leeds_houses,
  leeds_buildings,
  leeds_desire_lines,
  leeds_zones,
  output_format = "sf"
)
ab_list = ab_sf_to_json(ablines, mode_column = "mode_base")
ab_list$scenario
ab_list$people$trips[[1]]
dutch = ab_scenario(
  leeds_houses,
  leeds_buildings,
  leeds_desire_lines,
  leeds_zones,
  scenario = "godutch",
  output_format = "sf"
)
ab_list = ab_sf_to_json(dutch, mode_column = "mode_godutch")
ab_list$scenario
ab_list$people$trips[[9]]
# add times
dutch$departure = ab_time_normal(hr = 1, sd = 0, n = nrow(dutch))
ab_list_times = ab_sf_to_json(dutch)
f = tempfile(fileext = ".json")
ab_save(ab_list_times, f)
readLines(f)[13]
60^2
}
