% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{leeds_site_area}
\alias{leeds_site_area}
\alias{leeds_houses}
\alias{leeds_buildings}
\alias{leeds_desire_lines}
\alias{leeds_zones}
\title{Datasets from Leeds}
\description{
These datasets represent data from the case study city of Leeds.
}
